[20220618] 2번오류잡음

+++ src/main/java/community/controller/EventApplyUpdate.java
@@ -15,10 +15,8 @@ import com.oreilly.servlet.MultipartRequest;
 import com.oreilly.servlet.multipart.FileRenamePolicy;
 
 import common.HomeSweetHomeFileRenamePolicy;
-import community.model.dto.Attachment;
 import community.model.dto.EventAppAtt;
 import community.model.dto.EventAppExt;
-import community.model.dto.KnowhowExt;
 import community.model.service.EventAppService;
 
 /**
@@ -40,48 +38,53 @@ public class EventApplyUpdate extends HttpServlet {
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
+		
 		String saveDirectory = getServletContext().getRealPath("/upload/event");
 		int maxPostSize = 1024 * 1024 * 10;
 		String encoding = "utf-8";
 		FileRenamePolicy policy = new HomeSweetHomeFileRenamePolicy();
-		MultipartRequest multiReq = new MultipartRequest(request, saveDirectory, maxPostSize, encoding, policy);
-
+		MultipartRequest multiReq = 
+				new MultipartRequest(request, saveDirectory, maxPostSize, encoding, policy);
+	
+		
 		int no = Integer.parseInt(multiReq.getParameter("no"));
 		String eventapplyCode = multiReq.getParameter("eventapplyCode");
 		String memberId = multiReq.getParameter("memberId");
 		String nickName = multiReq.getParameter("nickName");
 		String content = multiReq.getParameter("content");
-		String[] delFiles = multiReq.getParameterValues("delFile");
-
+		String[] delFiles = multiReq.getParameterValues("delFile"); // 삭제하려는 첨부파일 pk
+	
 		EventAppExt eventapp = new EventAppExt();
 		eventapp.setNo(no);
 		eventapp.setEventapplyCode(eventapplyCode);
 		eventapp.setMemberId(memberId);
-		eventapp.setContent(content);
 		eventapp.setNickName(nickName);
+		eventapp.setContent(content);
+		
 
+		
 		File upFile1 = multiReq.getFile("upFile1");
 		File upFile2 = multiReq.getFile("upFile2");
-		if (upFile1 != null || upFile2 != null) {
+		if(upFile1 != null || upFile2 != null) {
 			List<EventAppAtt> attachments = new ArrayList<>();
-			if (upFile1 != null)
+			if(upFile1 != null)
 				attachments.add(getAttachment(multiReq, no, "upFile1"));
-			if (upFile2 != null)
+			if(upFile2 != null)
 				attachments.add(getAttachment(multiReq, no, "upFile2"));
 			eventapp.setAttachments(attachments);
 		}
-
+		
+		
 		int result = es.updateBoard(eventapp);
-
-		// 첨부파일 삭제
-		if (delFiles != null) {
-			for (String temp : delFiles) {
-				int attachNo = Integer.parseInt(temp);
+		
+		// 첨부파일 삭제 처리
+		if(delFiles != null) {
+			for(String temp : delFiles) {
+				int attachNo = Integer.parseInt(temp); // attachment pk
 				EventAppAtt attach = es.findAttachmentByNo(attachNo);
 
 				File delFile = new File(saveDirectory, attach.getRenamedFilename());
-				if (delFile.exists())
-					delFile.delete();
+				if(delFile.exists()) delFile.delete();
 
 				result = es.deleteAttachment(attachNo);
 				System.out.println("> " + attachNo + "번 첨부파일 (" + attach.getRenamedFilename() + ") 삭제!");
@@ -89,6 +92,7 @@ public class EventApplyUpdate extends HttpServlet {
 		}
 
 		response.sendRedirect(request.getContextPath() + "/event/eventApplyView?no=" + no);
+
 	}
 
 	private static EventAppAtt getAttachment(MultipartRequest multiReq, int no, String name) {
